refactor keycloak_group_roles implementation, add missing docs for keycloak_user_roles (#331)
