+++ keycloak/user_role_mappings.go
@@ -2,23 +2,8 @@ package keycloak
 
 import "fmt"
 
-// struct for the MappingRepresentation
-// https://www.keycloak.org/docs-api/8.0/rest-api/index.html#_mappingsrepresentation
-type UserRoleMapping struct {
-	ClientMappings map[string]*ClientRoleMapping `json:"clientMappings"`
-	RealmMappings  []*Role                       `json:"realmMappings"`
-}
-
-// struct for the ClientMappingRepresentation
-// https://www.keycloak.org/docs-api/8.0/rest-api/index.html#_clientmappingsrepresentation
-type ClientRoleMapping struct {
-	Client   string  `json:"client"`
-	Id       string  `json:"id"`
-	Mappings []*Role `json:"mappings"`
-}
-
-func (keycloakClient *KeycloakClient) GetUserRoleMappings(realmId string, userId string) (*UserRoleMapping, error) {
-	var roleMapping *UserRoleMapping
+func (keycloakClient *KeycloakClient) GetUserRoleMappings(realmId string, userId string) (*RoleMapping, error) {
+	var roleMapping *RoleMapping
 	err := keycloakClient.get(fmt.Sprintf("/realms/%s/users/%s/role-mappings", realmId, userId), &roleMapping, nil)
 	if err != nil {
 		return nil, err
