+++ mkdocs.yml
@@ -13,8 +13,9 @@ nav:
 - Resources:
   - keycloak_realm: resources/keycloak_realm.md
   - keycloak_realm_events: resources/keycloak_realm_events.md
-  - keycloak_user: resources/keycloak_user.md
   - keycloak_role: resources/keycloak_role.md
+  - keycloak_user: resources/keycloak_user.md
+  - keycloak_user_roles: resources/keycloak_user_roles.md
   - keycloak_group: resources/keycloak_group.md
   - keycloak_group_memberships: resources/keycloak_group_memberships.md
   - keycloak_group_roles: resources/keycloak_group_roles.md
